This reverts commit 1265.

+++ test/pthread2.c
@@ -43,8 +43,8 @@ static void *threadMain (void *data)
   int id = *(int*) data;
 
   while (1) {
-    printf("Worker: %d (%ld) alive. numWorkers: %d\n",
-           id, (long)syscall(SYS_gettid), numWorkers);
+    printf("Worker: %d (%d) alive. numWorkers: %d\n",
+           id, syscall(SYS_gettid), numWorkers);
     usleep(100*1000);
     pthread_mutex_lock(&mutex);
     if (numWorkers > 5) {
